add beginning of navigation and first modular css
