+++ package.json
@@ -18,7 +18,7 @@
     "react-bootstrap-pagination-control": "^1.0.5",
     "react-dom": "^18.2.0",
     "react-icons": "^4.8.0",
-    "react-router-dom": "^6.9.0",
+    "react-router-dom": "^6.10.0",
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
   },
