--- src/components/styles/admin-panel-styles.css
@@ -1,142 +1,0 @@
-.card {
-    height: auto;
-    width: 25rem;
-    box-shadow: 2px 4px 8px 0px rgb(0, 0, 0);
-    margin: 0rem auto auto auto;
-}
-
-.card:hover {
-    box-shadow: 2px 4px 8px 0px rgb(0, 0, 0);
-    transform: scale(1.0);
-}
-
-.modal-500w {
-    width: 470px;
-}
-
-.modal-800w {
-    width: 800px;
-}
-
-.default-button-radius {
-    border-radius: 12px;
-}
-
-.admin-panel-header {
-    padding: 6px;
-
-    padding-top: 15px;
-    background-color: #4169E1;
-    color: aliceblue;
-    width: 280px;
-    text-align: center;
-    border-radius: 0px 0px 24px 24px;
-    margin-left: auto;
-    margin-right: auto;
-    box-shadow: 0px 2px 4px 0px rgb(0, 0, 0);
-}
-
-.list-item-body {
-    margin-bottom: 20px;
-    margin-left: 10px;
-    margin-right: 10px;
-    padding: 10px;
-    border-radius: 20px;
-}
-
-.fab {
-    display: flex;
-    flex-direction: column-reverse;
-    list-style: none;
-    margin: 0;
-    padding: 0.8em;
-
-    border-radius: 100px;
-    position: fixed;
-    right: 2em;
-    bottom: 1.5em;
-}
-
-.password-visibility-button {
-    margin-left: 10px;
-    color: aliceblue;
-    display: inline-flex;
-    align-items: center;
-    justify-content: center;
-}
-
-.password-visibility-button:hover {
-    cursor: pointer;
-}
-
-.password-visibility-inline {
-    background-color: #4169E1;
-    border-radius: 0px 14px 14px 0px;
-    width: 40px;
-}
-
-.modal-content {
-    border-radius: 24px;
-    background-color: #252525;
-    color: aliceblue;
-}
-
-.modal-header {
-    border-bottom: 0 none;
-    padding-bottom: 0px;
-}
-
-.modal-footer {
-    border-top: 0 none;
-    padding: 0px 10px 10px 0px;
-}
-
-.form-margin-top {
-    margin-top: 10px;
-}
-
-.default-text-field {
-    background-color: #3b3a3a !important;
-    color: antiquewhite !important;
-    border-width: 0px;
-    border-color: transparent;
-    border-radius: 14px;
-}
-
-.default-text-field:hover {
-    background-color: #3b3a3a !important;
-    color: antiquewhite !important;
-}
-
-.default-text-field:focus {
-    background-color: #3b3a3a !important;
-    color: antiquewhite !important;
-    box-shadow: none;
-    border-width: 1px;
-    border-color: #4169E1;
-}
-
-.default-selector {
-    background-color: #3b3a3a !important;
-    color: antiquewhite !important;
-    border-width: 0px;
-    border-color: transparent;
-    border-radius: 14px;
-}
-
-.default-selector:focus {
-    box-shadow: none;
-    border-width: 1px;
-    border-color: #4169E1;
-}
-
-.one-line-text {
-    white-space: nowrap;
-    text-overflow: ellipsis;
-    color: red;
-    overflow: hidden;
-}
-
-body {
-    background-color: #252525;
-}
