+++ src/components/styles/adminPanel.module.css
@@ -1,0 +1,142 @@
+.card {
+    height: auto;
+    width: 30rem;
+    box-shadow: 2px 4px 8px 0px rgb(0, 0, 0);
+    margin: 0rem auto auto auto;
+}
+
+.modal500w {
+    width: 470px;
+    border-radius: 24px ;
+}
+
+.modal-800w {
+    width: 800px;
+}
+
+.defaultButtonRadius {
+    border-radius: 12px ;
+}
+
+.adminPanelHeader {
+    padding: 6px;
+    padding-top: 15px;
+    background-color: #4169E1;
+    color: aliceblue;
+    width: 280px;
+    text-align: center;
+    border-radius: 0px 0px 24px 24px;
+    margin-left: auto;
+    margin-right: auto;
+    box-shadow: 0px 2px 4px 0px rgb(0, 0, 0);
+}
+
+.userListItemBody {
+    margin-bottom: 20px;
+    margin-left: 10px;
+    margin-right: 10px;
+    padding: 10px;
+    border-radius: 20px ;
+    color: aliceblue;
+    background-color: #3b3a3a ;
+}
+
+.adminPanelFab {
+    display: flex;
+    flex-direction: column-reverse;
+    list-style: none;
+    margin: 0;
+    padding: 0.8em ;
+    border-radius: 100px ;
+    position: fixed;
+    right: 2em;
+    bottom: 1.5em;
+}
+
+.passwordVisibilityButton {
+    margin-left: 10px;
+    color: aliceblue;
+    display: inline-flex;
+    align-items: center;
+    justify-content: center;
+}
+
+.passwordVisibilityButton:hover {
+    cursor: pointer;
+}
+
+.passwordVisibilityInline {
+    background-color: #4169E1;
+    border-radius: 0px 14px 14px 0px;
+    width: 40px;
+}
+
+.adminPanelModalContent {
+    border-radius: 24px ;
+    background-color: #252525;
+    color: aliceblue;
+}
+
+.adminPanelModalHeader {
+    border-bottom: 0 none ;
+    padding-bottom: 0px;
+}
+
+.adminPanelModalFooter {
+    border-top: 0 none ;
+    padding: 0px 10px 10px 0px;
+}
+
+.formMarginTop {
+    margin-top: 10px ;
+}
+
+.defaultTextField {
+    background-color: #3b3a3a ;
+    color: antiquewhite ;
+    border-width: 0px ;
+    border-color: transparent ;
+    border-radius: 14px ;
+}
+
+.passRadius {
+    border-radius: 14px 0px 0px 14px ;
+}
+
+.defaultTextField:hover {
+    background-color: #3b3a3a ;
+    color: antiquewhite ;
+}
+
+.defaultTextField:focus {
+    background-color: #3b3a3a ;
+    color: antiquewhite ;
+    box-shadow: none;
+    border-width: 1px ;
+    border-color: #4169E1 ;
+}
+
+.defaultSelector {
+    background-color: #3b3a3a ;
+    color: antiquewhite ;
+    border-width: 0px ;
+    border-color: transparent ;
+    border-radius: 14px ;
+}
+
+.defaultSelector:focus {
+    box-shadow: none;
+    border-width: 1px ;
+    border-color: #4169E1 ;
+}
+
+.one-line-text {
+    white-space: nowrap;
+    text-overflow: ellipsis;
+    color: red;
+    overflow: hidden;
+}
+
+body {
+    background-color: #252525 ;
+}
