+++ src/components/styles/global-styles.css
@@ -23,3 +23,15 @@
     height: 100%;
     margin: auto;
 }
+
+.modal-content {
+    border-radius: 30px;
+}
+
+.modal-header {
+    border: 0px none;
+}
+
+.modal-footer {
+    border: 0px none;
+}
